+++ androidApp/app/src/main/java/it/unive/dais/legodroid/app/OnSensorChangedListener.java
@@ -1,6 +1,6 @@
 package it.unive.dais.legodroid.app;
 
 public interface OnSensorChangedListener {
-    void onUltrasonicChanged(float distance);
+    void onUltrasonicChanged(int distance);
     void onTouchSensorChanged(boolean isTouch);
 }
